Stop using StringUtils where not needed
The benefit of StringUtils::explode() is that it can work with
extremely large strings. This is not an issue here.

The benefit of StringUtils::explodeMarkup() is that it is able to
explode a string that contains HTML/XML tags, without destroying
these tags. This does not make sense here. Please refer to
https://www.mediawiki.org/wiki/Extension:ShowMe for an example.
This code parses the content of a <showme> tag. The result is going
to end as a list of options in an XmlSelect element. These options
must be plain text. HTML is not even allowed.

I don't think this qualifies as a breaking change. If there is a
<showme> tag out there with HTML elements, that's broken anyway
(the HTML is shown as plain text in the list of options).

Furthermore, how likely is it that the | character appears in a
nested HTML tag?

Other changes:
* Add missing trim() calls to both the label and the ID.
* Remove a not needed return true. This is the default for hook
  handler functions anyway.
* Replace deprecated addModuleScripts() that doesn't even exist any
  more, it seems.
* Update some PHPDoc type hints.

Change-Id: I1d04a98535d0ded4e81a2de8cfec2be84ee92b80